Add break-even lines and adjust margins in option graph
Enhanced the option graph visualization by adding break-even lines for both horizontal and vertical axes. Improved the range calculation by adjusting margins around the intrinsic value to ensure clarity. Changed sample option side to "Short" in the examples for testing purposes.

+++ examples/option_graph.rs
@@ -11,7 +11,7 @@ use std::error::Error;
 fn create_sample_option() -> Options {
     Options::new(
         OptionType::European,
-        Side::Long,
+        Side::Short,
         "AAPL".to_string(),
         100.0,
         ExpirationDate::Days(30.0),
